Experimenting on code-reuse
Trying to work out a way to define the invariants for vf:EconomicEvent, the
current development case is 'move'.
The idea is to have parametrizeable tests and to configure them for each
action. Eg. "'move' without a 'vf:toResourceInventoriedAs' does not affect
quantities, but it does must affect location. In this case, the first
part (selecting the instances of vf:EconomicEvent to test) is done with
a sh:SPARQLTargetFunction, the second part (testing the condition) is still unclear.

Two approaches work but are less than ideal:
1. using a sh:ConstraintComponent, we can have one big SPARQL query
validate the event (e.g., test that the location changes to the one in
the event).
	´- Upside:
		* We can report a custom error message (very helpful!)
		* The connection to the 'move' Shape takes 1 line -
  		This makes configuration and maintenance of the effects of actions
		VERY easy.
  	 - Downside:
	 	* There are many kinds of relationships to test, which
		differ mostly by the paths to get to the properties that
		have to be equal, or add up, or somesuch. The tests
		themselves, however, are always the same, and they need
		quite a lot of SPARQL code, which would be duplicated
		many times - a maintenance problem.
2. using one sh:expression in the Shape of the 'move' action for each
condition we want to test, plus a couple of custom SPARQLFunctions for
the tests
	- Upside:
		* Minimal code duplication
	- Downside:
		* Limited error messages (you cannot include variables
		in the message, i.e. you have one message constant per
		sh:expression)
		* The 'move' Shape is bloated - ca 10 lines per
		condition, nitty-gritty details all over the place in
		there. Lots of ways to create bugs that are hard to find

Will try 1. and maybe find a way to reduce code duplication by
externalizing SPARQL into sh:SPARQLFunction shapes.